+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Significantly reduced awaitable object size by dynamically allocating it.
 - Reduced memory footprint by removing preallocated awaitable objects.
 - Objects returned by a PyAwaitable object's `__await__` are now garbage collected (*i.e.*, they don't leak with rare circular references).
+- Removed limit on number of stored callbacks or values.
 
 ## [1.3.0] - 2024-10-26